+++ build-expat.sh
@@ -65,12 +65,14 @@ rm -rf "$EXPAT_DIR" &>/dev/null
 bzip2 -dk < "$EXPAT_TAR" | tar xf -
 cd "$EXPAT_DIR" || exit 1
 
-cp ../patch/expat.patch .
-patch -u -p0 < expat.patch
-echo ""
+if [[ -e ../patch/expat.patch ]]; then
+    cp ../patch/expat.patch .
+    patch -u -p0 < expat.patch
+    echo ""
+fi
 
 # Fix sys_lib_dlsearch_path_spec and keep the file time in the past
-../fix-config.sh
+cp -p ../fix-config.sh .; ./fix-config.sh
 
     PKG_CONFIG_PATH="${BUILD_PKGCONFIG[*]}" \
     CPPFLAGS="${BUILD_CPPFLAGS[*]}" \
@@ -99,6 +101,9 @@ then
     echo "Failed to build libexpat"
     exit 1
 fi
+
+# Fix flags in *.pc files
+cp -p ../fix-pc.sh .; ./fix-pc.sh
 
 echo "**********************"
 echo "Testing package"
